refactor: split MemberDetails component into Profile and Stats
- smaller components for easier maintenance

- refactor theme application
