+++ 12_redux/src/components/NoResults.js
@@ -6,10 +6,7 @@ const NoResults = () => {
 	const { theme, setTheme } = useContext(ThemeContext);
 
 	return (
-		<div
-			className="centered-div"
-			data-theme={theme === "light" ? "light" : "dark"}
-		>
+		<div className="centered-div" data-theme={theme}>
 			<h2>No results found for this search!</h2>
 		</div>
 	);
