+++ 9_10_custom_hook_and_context_api/src/App.js
@@ -24,10 +24,7 @@ const AppLayout = () => {
 
 	return (
 		<ThemeContext.Provider value={{ theme: theme, setTheme: setTheme }}>
-			<div
-				className="content"
-				data-theme={`${theme === "light" ? "light" : "dark"}`}
-			>
+			<div className="content" data-theme={theme}>
 				<Navbar />
 				<Outlet />
 			</div>
