+++ 9_10_custom_hook_and_context_api/src/components/Loading.js
@@ -6,10 +6,7 @@ const Loading = () => {
 	const { theme, setTheme } = useContext(ThemeContext);
 
 	return (
-		<div
-			className="centered-div"
-			data-theme={theme === "light" ? "light" : "dark"}
-		>
+		<div className="centered-div" data-theme={theme}>
 			<h2>Loading...</h2>
 		</div>
 	);
